requirements.txt
PyYAML>=5.4
google-cloud-firestore
numpy
pandas
schedule
requests
python-telegram-bot
websocket-client
python-dotenv









